# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AutoGame API
  version: 1.0.0
  description: AutoGame backend API specification (Spring Boot)
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
  /api/v1/players:
    post:
      tags: [Player]
      summary: Create player
      operationId: createPlayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlayerRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePlayer'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Player]
      summary: Get players
      operationId: getPlayers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePlayerList'
  /api/v1/players/{id}:
    get:
      tags: [Player]
      summary: Get player by id
      operationId: getPlayer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePlayer'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/items/{userId}:
    get:
      tags: [Item]
      summary: Get user inventory
      operationId: getUserItems
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserItemList'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/characters/{userId}:
    get:
      tags: [Character]
      summary: Get character stats
      operationId: getCharacterStats
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCharacterStat'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/characters/upgrade:
    post:
      tags: [Character]
      summary: Upgrade character stat
      operationId: upgradeCharacterStat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpgradeCharacterStatRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCharacterStat'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/daily-quests/{userId}:
    get:
      tags: [DailyQuest]
      summary: Get daily quest status
      operationId: getDailyQuestStatus
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDailyQuestStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/daily-quests/progress:
    post:
      tags: [DailyQuest]
      summary: Add daily quest progress
      operationId: addDailyQuestProgress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyQuestProgressRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDailyQuestStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/daily-quests/claim:
    post:
      tags: [DailyQuest]
      summary: Claim daily quest reward
      operationId: claimDailyQuestReward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyQuestClaimRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDailyQuestStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/dungeons/{dungeonId}/progress/{userId}:
    get:
      tags: [Dungeon]
      summary: Get dungeon progress
      operationId: getDungeonProgress
      parameters:
        - name: dungeonId
          in: path
          required: true
          schema:
            type: string
            example: dungeon1
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDungeonProgress'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/dungeons/{dungeonId}/progress:
    post:
      tags: [Dungeon]
      summary: Update dungeon progress
      operationId: updateDungeonProgress
      parameters:
        - name: dungeonId
          in: path
          required: true
          schema:
            type: string
            example: dungeon1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDungeonProgressRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDungeonProgress'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  responses:
    BadRequest:
      description: Validation or business rule error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: userId must not be null
      required: [code, message]
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
        error:
          $ref: '#/components/schemas/ApiError'
      required: [success, data, error]
    CreatePlayerRequest:
      type: object
      properties:
        nickname:
          type: string
          maxLength: 20
      required: [nickname]
    PlayerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        level:
          type: integer
          format: int64
        gold:
          type: integer
          format: int64
      required: [id, nickname, level, gold]
...
    UserItemResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        itemId:
          type: string
        itemName:
          type: string
        quantity:
          type: integer
          format: int64
        upgradeLevel:
          type: integer
          format: int32
        quality:
          type: string
        attackBonus:
          type: integer
          format: int32
        defenseBonus:
          type: integer
          format: int32
        hpBonus:
          type: integer
          format: int32
        mpBonus:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
      required: [userId, itemId, itemName, quantity, upgradeLevel, quality, attackBonus, defenseBonus, hpBonus, mpBonus, updatedAt]
...
    CharacterStatType:
      type: string
      enum: [ATTACK, DEFENSE, MAX_HP, MAX_MP]
    UpgradeCharacterStatRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        statType:
          $ref: '#/components/schemas/CharacterStatType'
      required: [userId, statType]
    CharacterStatResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        attack:
          type: integer
          format: int64
        defense:
          type: integer
          format: int64
        maxHp:
          type: integer
          format: int64
        maxMp:
          type: integer
          format: int64
        attackLevel:
          type: integer
          format: int32
        defenseLevel:
          type: integer
          format: int32
        hpLevel:
          type: integer
          format: int32
        mpLevel:
          type: integer
          format: int32
        nextAttackGoldCost:
          type: integer
          format: int64
        nextDefenseGoldCost:
          type: integer
          format: int64
        nextHpGemCost:
          type: integer
          format: int64
        nextMpGemCost:
          type: integer
          format: int64
      required:
        [
          userId,
          attack,
          defense,
          maxHp,
          maxMp,
          attackLevel,
          defenseLevel,
          hpLevel,
          mpLevel,
          nextAttackGoldCost,
          nextDefenseGoldCost,
          nextHpGemCost,
          nextMpGemCost
        ]
    DailyQuestProgressRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        dungeonKillDelta:
          type: integer
          format: int32
          minimum: 0
        goldEarnedDelta:
          type: integer
          format: int64
          minimum: 0
        statUpgradeDelta:
          type: integer
          format: int32
          minimum: 0
      required: [userId, dungeonKillDelta, goldEarnedDelta, statUpgradeDelta]
    DailyQuestClaimRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
      required: [userId]
    DailyQuestStatusResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        questDate:
          type: string
          format: date
        dungeonKillCount:
          type: integer
          format: int32
        targetKillCount:
          type: integer
          format: int32
        goldEarned:
          type: integer
          format: int64
        targetGoldEarned:
          type: integer
          format: int64
        statUpgradeCount:
          type: integer
          format: int32
        targetStatUpgradeCount:
          type: integer
          format: int32
        claimable:
          type: boolean
        claimed:
          type: boolean
        rewardGold:
          type: integer
          format: int64
        rewardGem:
          type: integer
          format: int64
      required:
        [
          userId,
          questDate,
          dungeonKillCount,
          targetKillCount,
          goldEarned,
          targetGoldEarned,
          statUpgradeCount,
          targetStatUpgradeCount,
          claimable,
          claimed,
          rewardGold,
          rewardGem
        ]
    UpdateDungeonProgressRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        currentWave:
          type: integer
          format: int32
          minimum: 1
        maxUnlockedWave:
          type: integer
          format: int32
          minimum: 1
      required: [userId, currentWave, maxUnlockedWave]
    DungeonProgressResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        dungeonId:
          type: string
        currentWave:
          type: integer
          format: int32
        maxUnlockedWave:
          type: integer
          format: int32
      required: [userId, dungeonId, currentWave, maxUnlockedWave]
    ApiResponsePlayer:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PlayerResponse'
        error:
          nullable: true
      required: [success, data, error]
    ApiResponsePlayerList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/PlayerResponse'
        error:
          nullable: true
      required: [success, data, error]
...
    ApiResponseUserItemList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserItemResponse'
        error:
          nullable: true
      required: [success, data, error]
...
    ApiResponseCharacterStat:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/CharacterStatResponse'
        error:
          nullable: true
      required: [success, data, error]
    ApiResponseDailyQuestStatus:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/DailyQuestStatusResponse'
        error:
          nullable: true
      required: [success, data, error]
    ApiResponseDungeonProgress:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/DungeonProgressResponse'
        error:
          nullable: true
      required: [success, data, error]